backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skillpulse-ingest"
version = "0.1.0"
description = "SkillPulse job ingestion backend"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31.0",
  "python-dateutil>=2.9.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

